CIS: Add "auto none" as a valid value for contain-intrinsic-size.
This change affects how we treat "contain-intrinsic-size: auto none".
Specifically, we treat it as a valid value that acts as "none" initially
and uses last remembered size when that's available.

I2P&S:
https://groups.google.com/a/chromium.org/g/blink-dev/c/mEZVe7MT88k/m/Xy6toLkhDAAJ
Note that this only adds the feature in the experimental state. I will
sent a separate CL to move to stable once there are approvals.

R=andruud@chromium.org, ikilpatrick@chromium.org

Bug: 1453733
Change-Id: I77e4750e2d259d1dee8489d744afa395aa23a2eb