[Canvas] Draw nothing when filling a line segment
Filling a path that is just a line should draw nothing. A recent
optimization that uses SkCanvas::drawLine instead of SkCanvas::drawPath
to draw line segments caused a regression because drawLine() always
strokes paths. This CL fixes the bu by exiting early when attempting to
fill a path that is just a line.

Bug: 1450607
Change-Id: I3532b85f6c6123d3a4ec189d5336c03f579684b9