Apply font-size-adjust to fallback fonts
Fallback fonts ignore font-size-adjust. That is because the primary
font's description is used to create fallback fonts. As a fallback font
can have different font metrics, we should adjust its size.

The subpixel rendering might affect the result on Windows. To make it
happy, we tweak the font size in font-size-adjust-012.html. With this
change, Windows gets a regression on font-size-adjust-010.html due to
unknown reasons. We will handle it in a separate ticket [1].

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1454612

Test: external/wpt/css/css-fonts/font-size-adjust-012.html
      external/wpt/css/css-fonts/font-size-adjust-metrics-override.html

Bug: 451346
Change-Id: Iab8f14c21fdf501931e62e77e9059737722a050f